requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-cosmos

pytest==7.4.0
pytest-cov==4.1.0
mock==4.0.3

Flask==2.3.0
flask-cors==4.0.0
gunicorn==21.2.0